bms_project/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.5.0
python-dotenv==1.0.1
requests==2.32.3
whitenoise==6.7.0
psycopg2-binary==2.9.9
gunicorn==22.0.0 